Add non-spatial tables to a rangeMapper project

Add any dataset to the project. The dataset is saved in a separate table inside the project and labelled as a bio table.

rmap_add_bio(con, x, ID, name)

# S4 method for rmapConnection,data.table,character,character
rmap_add_bio(con, x, ID, name)

# S4 method for rmapConnection,ANY,character,missing
rmap_add_bio(con, x, ID, name)

# S4 method for rmapConnection,data.frame,character,character
rmap_add_bio(con, x, ID, name)

# S4 method for rmapConnection,sf,character,character
rmap_add_bio(con, x, ID, name)

Arguments

con
a rangeMapper connection made with rmap_connect() .

x
an object inheriting from base::data.frame().

ID
character string. name of the ID column, usually species name.

name
output table name. If name is missing then name is the same as x.

Value

TRUE when the table is written to the project file, FALSE otherwise.

Details

The bio tables contain the data which is then mapped with rmap_save_map() at each canvas cell and/or data used to create subsets with rmap_save_subset(). If the bio table inherits from sf then the geometry is silently dropped and only the non-spatial data are imported.

Examples

con = rmap_connect()
rmap_add_ranges(con, wrens, 'sci_name')
rmap_add_bio(con, wrens, 'sci_name')
dbDisconnect(con)